Made ndpi memory allocator type compatible with C++ on x686 platforms. Switch from unsigned long to size_t as allocation functions parameter.

+++ example/ndpiReader.c
@@ -407,7 +407,7 @@ static void debug_printf(u_int32_t protocol, void *id_struct,
 
 /* ***************************************************** */
 
-static void *malloc_wrapper(unsigned long size) {
+static void *malloc_wrapper(size_t size) {
   current_ndpi_memory += size;
 
   if(current_ndpi_memory > max_ndpi_memory)
